requirements.txt
alembic==1.13.2
altair==5.4.0
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
black==24.8.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
fastapi==0.115.12
flake8==7.1.1
folder-structure-generator==0.0.2
fonttools==4.53.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.2.3
h11==0.16.0
idna==3.7
iniconfig==2.1.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mccabe==0.7.0
mdurl==0.1.2
mypy-extensions==1.0.0
narwhals==1.33.0
numpy==2.2.4
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
playwright==1.52.0
plotly==6.0.1
pluggy==1.6.0
protobuf==5.29.5
psycopg2-binary==2.9.9
pyarrow==17.0.0
pycodestyle==2.12.1
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pyee==13.0.0
pyflakes==3.2.0
Pygments==2.18.0
pyparsing==3.1.4
pytest==8.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.18
pytz==2024.1
referencing==0.35.1
requests>=2.32.4
rich==13.7.1
rpds-py==0.20.0
rptree==0.1.1
scipy==1.14.0
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.32
st-theme==1.2.3
starlette==0.46.2
streamlit==1.45.0
streamlit-community-navigation-bar==4.0.9
tenacity==8.5.0
toml==0.10.2
tornado==6.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.34.2
watchdog==4.0.2
